Create Computational_Problems.md
This file will contain problems that students can try solving related to using python scripts for computational purposes. These problems are not intended to be directly relevant to the average professional software developer job. The students will add their solutions to their branch and make a PR and comments will be made to their PRs on the quality of the code written. The focus of these questions is to improve coding proficiency, rather than necessarily understanding the real-world solutions to the scenarios the scripts are being written for, although the real world component won't be completely neglected either.

Signed-off-by: Jack Zhang <89618353+zhanjack822@users.noreply.github.com>